+++ domain/src/main/java/daily/dayo/domain/usecase/post/RequestEditPostUseCase.kt
@@ -11,7 +11,7 @@ class RequestEditPostUseCase @Inject constructor(
         postId: Long,
         category: Category,
         contents: String,
-        folderId: Int,
+        folderId: Long,
         hashtags: List<String>
     ) =
         postRepository.requestEditPost(postId, category, contents, folderId, hashtags)
